Ai_Mawid/mawid-assistant/requirements.txt
# === Web Framework ===
fastapi
uvicorn

# === Supabase ===
supabase

# === RAG / Embeddings ===
sentence-transformers

# === Utilities ===
python-dotenv
pydantic
httpx
python-jose[cryptography]
